+++ src/index.js
@@ -1,6 +1,7 @@
-import Locator, { locator, } from './Locator';
-export { default as Locator, } from './Locator';
+import Locator, { locator, } from './Locator'
+export { default as Locator, } from './Locator'
+export const { get, config, register, clear, } = locator
 
-locator.Locator = Locator;
+locator.Locator = Locator
 
-export default locator;
+export default locator